+++ docs/CHANGE_LOG.md
@@ -2,6 +2,8 @@
 
 ## Version 1.9.2:
 - Updated the README documentation.
+- Added a "visible" property to the binding option configuration "mapRangeColors", which will allow you to set the visible ranges on load.
+- 
 
 <br>